drm/rockchip: vop2: fix esmart pd turn off timeout issue
The esmart pd attach to esmart1/esmart2/esmart3, if we want to
turn off esmart pd, a vp attach to esmart1, esmart2 or esmart3
must active.

Now, the esmart pd is turn off when the last vp is disable, if
the last vp is not attach to esmart1, esmart2 or esmart3, the
esmart pd will turn off failed.

To fix this issue, when disable the last vp, we will active the
vp attach emsart1 and turn off the esmart pd.

Signed-off-by: Zhang Yubing <yubing.zhang@rock-chips.com>
Change-Id: Ide8a367de2c8b7ba027568bc7eacb3fd389c73fe
Signed-off-by: Stephen Chen <stephen@radxa.com>